+++ modules/apps/fragment/fragment-renderer-react-impl/src/main/java/com/liferay/fragment/renderer/react/internal/model/listener/FragmentEntryLinkModelListener.java
@@ -15,7 +15,7 @@
 package com.liferay.fragment.renderer.react.internal.model.listener;
 
 import com.liferay.fragment.model.FragmentEntryLink;
-import com.liferay.fragment.renderer.react.internal.util.FragmentEntryFragmentRendererReactHelper;
+import com.liferay.fragment.renderer.react.internal.helper.FragmentEntryLinkJSModuleHelper;
 import com.liferay.portal.kernel.model.BaseModelListener;
 import com.liferay.portal.kernel.model.ModelListener;
 
@@ -37,8 +37,7 @@ public class FragmentEntryLinkModelListener
 			return;
 		}
 
-		_fragmentEntryFragmentRendererReactHelper.registerJSModule(
-			fragmentEntryLink);
+		_fragmentEntryLinkJSModuleHelper.registerJSModule(fragmentEntryLink);
 	}
 
 	@Override
@@ -47,10 +46,9 @@ public class FragmentEntryLinkModelListener
 			return;
 		}
 
-		_fragmentEntryFragmentRendererReactHelper.ensureInitialized();
+		_fragmentEntryLinkJSModuleHelper.ensureInitialized();
 
-		_fragmentEntryFragmentRendererReactHelper.unregisterJSModule(
-			fragmentEntryLink);
+		_fragmentEntryLinkJSModuleHelper.unregisterJSModule(fragmentEntryLink);
 	}
 
 	@Override
@@ -62,14 +60,13 @@ public class FragmentEntryLinkModelListener
 			return;
 		}
 
-		_fragmentEntryFragmentRendererReactHelper.ensureInitialized();
+		_fragmentEntryLinkJSModuleHelper.ensureInitialized();
 
-		_fragmentEntryFragmentRendererReactHelper.updateJSModule(
+		_fragmentEntryLinkJSModuleHelper.updateJSModule(
 			fragmentEntryLink, originalFragmentEntryLink);
 	}
 
 	@Reference
-	private FragmentEntryFragmentRendererReactHelper
-		_fragmentEntryFragmentRendererReactHelper;
+	private FragmentEntryLinkJSModuleHelper _fragmentEntryLinkJSModuleHelper;
 
 }
